+++ Simulation/Simulation/Brick/Odd/NoneUp.cs
@@ -10,8 +10,8 @@ namespace Simulation
 {
     class NoneUp : Brick
     {
-        private float probability;
-		private float startProbability;
+        private double probability;
+		private double startProbability;
 
 		private readonly bool upBond;
 		private readonly bool downBond;
@@ -141,7 +141,7 @@ namespace Simulation
             right = bricks[15];
         }
 
-        public float Probability { get { return probability; } set { probability = value; } }
+		public double Probability { get { return probability; } set { probability = value; } }
 
         public Color CouplingColor { get { return Color.Empty; } set { } }
 
@@ -151,7 +151,7 @@ namespace Simulation
 
         public TextBox Display { set { } }
 
-		public float StartProbability { get { return startProbability; } set { startProbability = value; } }
+		public double StartProbability { get { return startProbability; } set { startProbability = value; } }
 
 		public bool UpBond { get { return upBond; } }
